AVRO-4064: Move relavent unit tests from ipc package to avro package
